test(fuzza): split the differential test registry into thematic modules
tests.rs had grown to a flat 1051-line file of 99 test registrations,
which made review and navigation unwieldy and scattered related tests —
bug reproducers ended up far from the sibling cases that bound them, and
pinned-grid companions far from their base cases — because entries were
appended in campaign order.

Move the registrations verbatim into nine thematic modules under
tests/ (arith, control_flow, calls, memory, spills, signed, wide,
boundaries, scale), keeping every doc comment and ignore attribute
unchanged and each _repro/_edges companion directly below its base case;
include_str paths gain one level for the new directory depth. Bug
families are now physically adjacent: the sext_shapes pair sits with its
bounding siblings in signed, the switch pair in control_flow. The
module doc in tests/mod.rs states the filing rule, and the fuzza-agent
docs now point at the module directory with an updated wiring snippet.
